# docs/swagger.yaml
definitions:
  interfaces.AssetsInventory:
    properties:
      business_value:
        type: integer
      criticality:
        type: string
      description:
        type: string
      id:
        type: integer
      location:
        type: string
      name:
        type: string
      replacement_cost:
        type: number
      responsible:
        type: string
      target_environment:
        type: string
      users:
        type: string
    type: object
  interfaces.ControlImplementation:
    properties:
      controlId:
        type: integer
      current_implementation:
        type: integer
      current_percent_value:
        type: string
      id:
        type: integer
      projected_cost:
        type: number
      proposed_implementation:
        type: integer
      proposed_percent_value:
        type: string
    type: object
  interfaces.ControlLibrary:
    properties:
      control_reference:
        type: string
      control_type:
        type: string
      controlId:
        type: integer
      id:
        type: integer
      in_scope:
        type: string
      information:
        type: string
    type: object
...
  interfaces.InputControlImplementation:
    properties:
      controlId:
        type: integer
      current_implementation:
        type: integer
      current_percent_value:
        type: string
      projected_cost:
        type: number
      proposed_implementation:
        type: integer
      proposed_percent_value:
        type: string
    type: object
  interfaces.InputControlLibrary:
    properties:
      control_reference:
        type: string
      control_type:
        type: string
      controlId:
        type: integer
      in_scope:
        type: string
      information:
        type: string
    type: object
  interfaces.InputControlls:
    properties:
      airborne_particles_event:
        type: integer
      authentication_attack:
        type: integer
      authorisation_attack:
        type: integer
      biological_event:
        type: integer
      communication_attack:
        type: integer
      controlId:
        type: integer
      controlType:
        type: string
      denial_of_service_attack:
        type: integer
      geological_event:
        type: integer
      hydrological_event:
        type: integer
      information_leakage_attack:
        type: integer
      infrastructure_failure_event:
        type: integer
      malware_attack:
        type: integer
      meteorological_event:
        type: integer
      misconfiguration_attack:
        type: integer
      misuse_attack:
        type: integer
      natural_hazard_event:
        type: integer
      people_failure:
        type: integer
      physical_attack:
        type: integer
      process_failure:
        type: integer
      reconnaissance_activities:
        type: integer
      social_engineering_attack:
        type: integer
      software_exploitation_attack:
        type: integer
      supply_chain_attack:
        type: integer
      technology_failure:
        type: integer
    type: object
...
  interfaces.InputThreatEventCatalogue:
    properties:
      description:
        type: string
      id:
        type: integer
      in_scope:
        type: string
      threat_event:
        type: string
      threat_group:
        type: string
    type: object
...
  interfaces.Propused:
    properties:
      airborne_particles_event:
        type: integer
      authentication_attack:
        type: integer
      authorisation_attack:
        type: integer
      biological_event:
        type: integer
      communication_attack:
        type: integer
      controlId:
        type: integer
      denial_of_service_attack:
        type: integer
      geological_event:
        type: integer
      hydrological_event:
        type: integer
      id:
        type: integer
      information_leakage_attack:
        type: integer
      infrastructure_failure_event:
        type: integer
      malware_attack:
        type: integer
      meteorological_event:
        type: integer
      misconfiguration_attack:
        type: integer
      misuse_attack:
        type: integer
      natural_hazard_event:
        type: integer
      people_failure:
        type: integer
      physical_attack:
        type: integer
      process_failure:
        type: integer
      reconnaissance_activities:
        type: integer
      social_engineering_attack:
        type: integer
      software_exploitation_attack:
        type: integer
      supply_chain_attack:
        type: integer
      technology_failure:
        type: integer
    type: object
  interfaces.Relevance:
    properties:
      airborne_particles_event:
        type: integer
      authentication_attack:
        type: integer
      authorisation_attack:
        type: integer
      biological_event:
        type: integer
      communication_attack:
        type: integer
      denial_of_service_attack:
        type: integer
      geological_event:
        type: integer
      hydrological_event:
        type: integer
      id:
        type: integer
      information_leakage_attack:
        type: integer
      infrastructure_failure_event:
        type: integer
      malware_attack:
        type: integer
      meteorological_event:
        type: integer
      misconfiguration_attack:
        type: integer
      misuse_attack:
        type: integer
      natural_hazard_event:
        type: integer
      people_failure:
        type: integer
      physical_attack:
        type: integer
      process_failure:
        type: integer
      reconnaissance_activities:
        type: integer
      social_engineering_attack:
        type: integer
      software_exploitation_attack:
        type: integer
      supply_chain_attack:
        type: integer
      technology_failure:
        type: integer
    type: object
...
  interfaces.Strength:
    properties:
      airborne_particles_event:
        type: integer
      authentication_attack:
        type: integer
      authorisation_attack:
        type: integer
      biological_event:
        type: integer
      communication_attack:
        type: integer
      controlId:
        type: integer
      denial_of_service_attack:
        type: integer
      geological_event:
        type: integer
      hydrological_event:
        type: integer
      id:
        type: integer
      information_leakage_attack:
        type: integer
      infrastructure_failure_event:
        type: integer
      malware_attack:
        type: integer
      meteorological_event:
        type: integer
      misconfiguration_attack:
        type: integer
      misuse_attack:
        type: integer
      natural_hazard_event:
        type: integer
      people_failure:
        type: integer
      physical_attack:
        type: integer
      process_failure:
        type: integer
      reconnaissance_activities:
        type: integer
      social_engineering_attack:
        type: integer
      software_exploitation_attack:
        type: integer
      supply_chain_attack:
        type: integer
      technology_failure:
        type: integer
    type: object
...
  interfaces.ThreatEventCatalogue:
    properties:
      description:
        type: string
      id:
        type: integer
      in_scope:
        type: string
      threat_event:
        type: string
      threat_group:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is a server for app.
  title: Qira
  version: "1.0"
paths:
...
  /api/all-catalogue:
    get:
      consumes:
      - application/json
      description: Retrieve all Event
      parameters:
      - default: Bearer <token>
        description: Auth Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of All Event
          schema:
            $ref: '#/definitions/interfaces.ThreatEventCatalogue'
      summary: Retrieve All Catalogue
      tags:
      - Catalogue
...
  /api/catalogue:
    post:
      consumes:
      - application/json
      description: Create New Event Catalogue
      parameters:
      - description: Data for create new Event
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/interfaces.InputThreatEventCatalogue'
      - default: Bearer <token>
        description: Auth Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of All Assets
          schema:
            $ref: '#/definitions/interfaces.ThreatEventCatalogue'
      summary: Create New Catalogue
      tags:
      - Catalogue
  /api/catalogue/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve an Event by its ID
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: integer
      - default: Bearer <token>
        description: Auth Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Event Details
          schema:
            $ref: '#/definitions/interfaces.AssetsInventory'
      summary: Retrieve Catalogue by ID
      tags:
      - Catalogue
...
  /api/create-implementation:
    post:
      consumes:
      - application/json
      description: Create new Control Implementation
      parameters:
      - description: Data for create new Risk
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/interfaces.InputControlImplementation'
      - default: Bearer <token>
        description: Auth Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Risk Create
          schema:
            $ref: '#/definitions/interfaces.ControlImplementation'
      summary: Create Control Implementation
      tags:
      - Control
  /api/create-library:
    post:
      consumes:
      - application/json
      description: Create new Control Library
      parameters:
      - description: Data for create new Risk
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/interfaces.InputControlLibrary'
      - default: Bearer <token>
        description: Auth Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Risk Create
          schema:
            $ref: '#/definitions/interfaces.ControlLibrary'
      summary: Create Control Library
      tags:
      - Control
...
  /api/create-propused:
    post:
      consumes:
      - application/json
      description: Create new Control Propused
      parameters:
      - description: Data for create new Risk
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/interfaces.InputControlls'
      - default: Bearer <token>
        description: Auth Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Risk Create
          schema:
            $ref: '#/definitions/interfaces.Propused'
      summary: Create Control Propused
      tags:
      - Control
  /api/create-relevance:
    post:
      consumes:
      - application/json
      description: Create new Control Relevance
      parameters:
      - description: Data for create new Risk
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/interfaces.InputControlls'
      - default: Bearer <token>
        description: Auth Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Risk Create
          schema:
            $ref: '#/definitions/interfaces.Relevance'
      summary: Create Control Relevance
      tags:
      - Control
  /api/create-strength:
    post:
      consumes:
      - application/json
      description: Create new Control Strength
      parameters:
      - description: Data for create new Risk
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/interfaces.InputControlls'
      - default: Bearer <token>
        description: Auth Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Risk Create
          schema:
            $ref: '#/definitions/interfaces.Strength'
      summary: Create Control Strength
      tags:
      - Control
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"